refac: file structure refactored
